remove CR step 2, 3,4,use one file
